kafka_producer/requirements.txt
kafka-python
tweepy